Refactor support functions into modules
