+++ support/__init__.py
@@ -1,4 +1,6 @@
-# __init__.py
-
-from .functions import *
+# Package shortcut imports
+from .image_utils import *
+from .file_utils import *
+from .style_utils import *
+from .story_utils import *
 from .logger import *
